[Refactoring] Consolidate HTTP/SSE functionality into main index.ts
- Remove separate http.ts file
- Merge HTTP/SSE server transport functionality into index.ts
- Update Dockerfile to use index.js as entrypoint instead of http.js
- Maintain all existing HTTP/SSE capabilities in unified structure
